requirements.txt
Django>=4.2,<6.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
markdown>=3.5.0
django-markdownify>=0.9.3
Pillow>=10.0.0
django-filter>=24.1
